AR-700: Refactored modification traits to base class to use lifecycle call backs
