+++ src/Entity/Tenant/Media.php
@@ -3,13 +3,11 @@
 namespace App\Entity\Tenant;
 
 use App\Entity\EntityIdTrait;
-use App\Entity\EntityModificationTrait;
 use App\Entity\EntityTitleDescriptionTrait;
 use App\Repository\MediaRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Timestampable\Traits\TimestampableEntity;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Validator\Constraints as Assert;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
@@ -22,9 +20,7 @@ use Vich\UploaderBundle\Mapping\Annotation as Vich;
 class Media extends AbstractTenantScopedEntityScoped
 {
     use EntityIdTrait;
-    use TimestampableEntity;
     use EntityTitleDescriptionTrait;
-    use EntityModificationTrait;
 
     /**
      * @Vich\UploadableField(mapping="media_object", fileNameProperty="filePath")
