+++ src/Entity/Tenant/Theme.php
@@ -3,13 +3,11 @@
 namespace App\Entity\Tenant;
 
 use App\Entity\EntityIdTrait;
-use App\Entity\EntityModificationTrait;
 use App\Entity\EntityTitleDescriptionTrait;
 use App\Repository\ThemeRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Timestampable\Traits\TimestampableEntity;
 
 /**
  * @ORM\Entity(repositoryClass=ThemeRepository::class)
@@ -19,8 +17,6 @@ class Theme extends AbstractTenantScopedEntityScoped
 {
     use EntityIdTrait;
     use EntityTitleDescriptionTrait;
-    use EntityModificationTrait;
-    use TimestampableEntity;
 
     /**
      * @ORM\Column(type="text")
